Make it possible to test runners that don't support all metrics
Runners may only partially support metrics. This partial support runs along axes:
- attempted & committed
- counters & distributions & gauges & ...

Prior to this PR, we have categories for the first axis, but not the second.
This means that a runner that supports only part of the second axis has to blacklist tests
on the first axis. Adding categories for both axes lets runners build a matrix of supported
features.

(This also lets the DataflowRunner run more tests by accurately identifying its test
 matrix.)

* MetricsMatchers: refactor matchers for committed/attempted code reuse